fix(contracts): keep date hygiene off the CONFENGE freeze tree (#552)
Revert scripts/contracts_truth.py (protected input). Persist-time
quarantine and status_observed_at stay in the SQL trigger; the Python
helper lives in date_semantics.py.

(cherry picked from commit 1a0b99a64dae224a725e5a9a4be62e453b3eaf7a)
